bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
aiosqlite>=0.19.0
